Address slash completion review feedback
